Refactor Add and remove book
