+++ src/components/pages/Categories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { checkStatus } from '../../redux/categories/categories';
+import checkStatus from '../../redux/actions/categories';
 
 function Categories() {
   const dispatch = useDispatch();
